rename: util 관련 파일 이름 케밥케이스로 통일
