+++ app/feature/auth/utils/encrypt-data.ts
