+++ app/feature/search/queries/postKeywordSearchCurationData.ts
@@ -1,4 +1,4 @@
-import { getSession } from "@/common/utils/session/getSession";
+import { getSession } from "@/common/utils/session/get-session";
 
 export async function postKeywordSearchCurationData(keyword: any) {
   const auth_info = await getSession();
