+++ app/feature/user/queries/getMemberInfo.tsx
@@ -1,5 +1,5 @@
 import ApiErrorMessage from "@/common/utils/error/api-error-message";
-import { getSession } from "@/common/utils/session/getSession";
+import { getSession } from "@/common/utils/session/get-session";
 
 export default async function GetMemberInfo(): Promise<{
   nickname: string;
